14_Documentation/djlibrary/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
config==0.5.0.post0
coreapi==2.3.3
coreschema==0.0.4
databases==0.5.2
distlib==0.3.2
Django==4.0
django-modeltranslation==0.17.3
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
docutils==0.18.1
drf-yasg==1.20.0
filelock==3.0.12
filesystems==0.26.0
Flask==1.1.2
greenlet==1.1.2
homoglyphs==2.0.4
idna==3.2
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.2
loguru==0.5.3
MarkupSafe==2.0.1
packaging==21.0
pathlib==1.0.1
phonenumbers==8.12.35
Pillow==8.4.0
platformdirs==2.3.0
pygame==2.0.1
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.18.0
pyTelegramBotAPI==4.0.0
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
sqlparse==0.4.1
toml==0.10.2
tomlkit==0.7.2
tqdm==4.62.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.6
venvs==7.1.2
virtualenv==20.7.2
Werkzeug==2.0.2
win32-setctime==1.0.3
zope.interface==5.4.0
